Move the base classes for lambdas into the common project
