+++ monthly-contributions/src/main/scala/com/gu/support/workers/lambdas/CreateZuoraSubscription.scala
@@ -1,7 +1,6 @@
 package com.gu.support.workers.lambdas
 
 import com.amazonaws.services.lambda.runtime.Context
-import com.gu.support.workers.helpers.Handler
 import com.gu.support.workers.model.SalesForceContact
 import com.typesafe.scalalogging.LazyLogging
 import io.circe.generic.auto._
